Update amazon-glacier-data-model.md
On line #71 "jobs" should be singular because the description of the actions are singular. If "jobs" is kept plural then the phrases "select query on an archive, retrieve an archive, or get an inventory of a vault" should be made plural. 

Thank you for taking the time to review my proposed change.