Update amazon-glacier-getting-started.md
On line #3 the word "the" is awkward. It implies specific vaults and resources rather than generic terms. 
The opening paragraph describes what vaults and archives are so mentioning that they are describe in the Data Model section is a little redundant. It is a little better to instruct the reader to seek additional information. 

Thank you for taking the time to review my proposed change.